refactor: extract AXIS vector storage to proximadb-index-storage foundation crate
Move src/index/axis/{zero_overhead_vector,compact_vector}.rs (~840 LOC) into a
new foundation crate `proximadb-index-storage`. The root keeps pure
`pub use` re-export shims so every `crate::index::axis::{zero_overhead,compact}_vector::*`
path resolves unchanged — no caller edits.

Second increment of the shared hermetic-leaf unlock in the root-crate
decomposition track. The two vector representation files (ZeroOverheadVector,
CompactVector + their collections) are hermetic (zero crate:: imports; extern
deps only: anyhow, bytemuck, dashmap) and are the next most-depended-on
hermetic root-internal pair after concurrent_structures (#1071): structural
edges from the ORION graph engine and the pgwire layer. Extracting them to a
leaf removes another recurring root-internal dependency that would force those
cohorts into a cyclic root dependency when extracted.

Pure relocation + re-export, no behavior change. Crate-level clippy allows
mirror the root crate (code written under those suppressions).
