+++ crates/foundation/proximadb-index-storage/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "proximadb-index-storage"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+authors.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "ProximaDB foundation-tier AXIS index vector storage representations (ZeroOverheadVector, CompactVector) — concrete packed/zero-copy storage for index implementations, shared by storage engines and index layers"
+
+[lib]
+name = "proximadb_index_storage"
+path = "src/lib.rs"
+
+[dependencies]
+anyhow.workspace = true
+bytemuck = { version = "1.14", features = ["derive"] }
+dashmap.workspace = true
