+++ src/index/axis/zero_overhead_vector.rs
@@ -1,472 +1,13 @@
-/*
- * Copyright 2025 ProximaDB
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- */
-
-//! Zero-Overhead Vector Storage for AXIS Indexes
+// Copyright (C) 2025 ProximaDB
+// SPDX-License-Identifier: Apache-2.0
+//! Re-export shim.
 //!
-//! # The Zero-Overhead Promise
-//!
-//! Traditional vector databases waste 50-90% of memory on metadata that never changes.
-//! ProximaDB's Zero-Overhead Vector Storage eliminates ALL redundant metadata by
-//! leveraging collection-level configuration cached in shared memory.
-//!
-//! ## What is Zero-Overhead?
-//!
-//! - **Zero metadata per vector** - no dimension, no flags, no repeated config
-//! - **Zero serialization cost** - direct bytemuck memory mapping
-//! - **Zero-copy access** - SIMD-ready aligned memory
-//! - **Zero wasted bytes** - only store what's unique: vector data + ID
-//!
-//! # Design Evolution: From 1600 Bytes to 400 Bytes
-//!
-//! ## The Problem: Traditional VectorRecord
-//! ```text
-//! VectorRecord {
-//!     id: String,           // ✓ Needed
-//!     vector: Vec<f32>,     // ✓ Needed  
-//!     metadata: HashMap,    // ✗ Not needed for similarity search
-//!     timestamp: u64,       // ✗ Not needed for similarity search
-//!     version: u32,         // ✗ Not needed for similarity search
-//!     expires_at: u64,      // ✗ Not needed for similarity search
-//!     ... more fields ...   // ✗ Not needed for similarity search
-//! }
-//! Size: ~1600 bytes for 384-dim vector
-//! ```
-//!
-//! ## The Journey to Zero
-//!
-//! ### Insight #1: "Why store full records when we only need vectors?"
-//! Eliminated proto overhead → Saved 3%
-//!
-//! ### Insight #2: "Why store ID length when it's deducible?"
-//! After fixed-size vector, everything else is the ID → Saved another 2 bytes
-//!
-//! ### Insight #3: "Why store dimension when it's constant per collection?"
-//! Dimension lives in collection config (O(1) cache) → Saved 2 bytes per vector
-//!
-//! ### Insight #4: "Why store quantization info when it's also constant?"
-//! Quantization config is per-collection → Saved final byte of overhead
-//!
-//! ## The Result: True Zero-Overhead Storage
-//!
-//! ```text
-//! ZeroOverheadVector {
-//!     data: [vector_bytes][id_string]  // That's it. Nothing else.
-//! }
-//! ```
-//!
-//! ### Memory Savings by Vector Type:
-//! - **FP32 (384-dim)**: 1600 → 1552 bytes (3% saved, but zero metadata!)
-//! - **INT8 (384-dim)**: 1600 → 400 bytes (75% saved!)
-//! - **PQ4 (384-dim)**: 1600 → 208 bytes (87% saved!)
-//! - **Binary (384-dim)**: 1600 → 64 bytes (96% saved!)
-//!
-//! # Why Not Compress Further?
-//!
-//! **Q: Should we compress the raw vectors?**
-//! **A: No!** Here's why:
-//!
-//! 1. **Poor ROI**: Dense embeddings compress only 1.1-1.3x (random floats)
-//! 2. **Performance Hit**: Loses zero-copy, breaks SIMD alignment (10-100x slower)
-//! 3. **Better Alternative**: Quantization already gives 4-32x reduction
-//! 4. **Cache Pollution**: Decompression buffers thrash CPU cache
-//!
-//! # Implementation
-
-use anyhow::{Result, anyhow};
-use bytemuck;
-use dashmap::DashMap;
-use std::sync::Arc;
-
-/// Zero-overhead vector storage - absolutely minimal memory footprint
-///
-/// Layout: [vector_data][id_string]
-/// No headers. No metadata. Just data.
-#[derive(Debug, Clone)]
-pub struct ZeroOverheadVector {
-    data: Vec<u8>,
-}
-
-impl ZeroOverheadVector {
-    /// Create from FP32 vector - zero overhead
-    #[inline]
-    pub fn from_fp32(id: &str, vector: &[f32]) -> Result<Self> {
-        let vector_bytes = bytemuck::cast_slice::<f32, u8>(vector);
-        let id_bytes = id.as_bytes();
-
-        let mut data = Vec::with_capacity(vector_bytes.len() + id_bytes.len());
-        data.extend_from_slice(vector_bytes);
-        data.extend_from_slice(id_bytes);
-
-        Ok(Self { data })
-    }
-
-    /// Create from quantized vector - zero overhead
-    #[inline]
-    pub fn from_quantized(id: &str, quantized_vector: &[u8]) -> Result<Self> {
-        let id_bytes = id.as_bytes();
-
-        let mut data = Vec::with_capacity(quantized_vector.len() + id_bytes.len());
-        data.extend_from_slice(quantized_vector);
-        data.extend_from_slice(id_bytes);
-
-        Ok(Self { data })
-    }
-
-    /// Get vector as FP32 (zero-copy) - requires dimension from collection
-    #[inline]
-    pub fn as_f32(&self, dimension: usize) -> &[f32] {
-        let vector_bytes = dimension * std::mem::size_of::<f32>();
-        let bytes = &self.data[..vector_bytes];
-        bytemuck::cast_slice(bytes)
-    }
-
-    /// Get vector as quantized bytes - requires size from collection config
-    #[inline]
-    pub fn as_quantized(&self, quantized_size: usize) -> &[u8] {
-        &self.data[..quantized_size]
-    }
-
-    /// Get ID - requires vector size to know where ID starts
-    #[inline]
-    pub fn id(&self, vector_size: usize) -> &str {
-        let id_bytes = &self.data[vector_size..];
-        // SAFETY: We only store valid UTF-8 strings
-        unsafe { std::str::from_utf8_unchecked(id_bytes) }
-    }
-
-    /// Total size in bytes
-    #[inline]
-    pub fn size_bytes(&self) -> usize {
-        self.data.len()
-    }
-
-    /// Get raw data for serialization
-    #[inline]
-    pub fn as_bytes(&self) -> &[u8] {
-        &self.data
-    }
-
-    /// Create from raw bytes for deserialization
-    #[inline]
-    pub fn from_bytes(data: Vec<u8>) -> Self {
-        Self { data }
-    }
-}
-
-/// Backwards-compat alias for [`ZeroOverheadCollectionConfig`].
-pub type CollectionConfig = ZeroOverheadCollectionConfig;
-
-/// Configuration for a zero-overhead collection
-#[derive(Debug, Clone)]
-pub struct ZeroOverheadCollectionConfig {
-    /// Number of dimensions in each vector
-    pub dimension: usize,
-    /// Whether vectors are stored in quantized form
-    pub is_quantized: bool,
-    /// Quantization scheme applied to vectors, if any
-    pub quantization_method: Option<QuantizationMethod>,
-}
-
-/// Quantization method applied to stored vectors
-#[derive(Debug, Clone, Copy)]
-pub enum QuantizationMethod {
-    /// Scalar quantization: 1 byte per dimension
-    INT8,
-    /// Product quantization with 1 byte per sub-space (PQ-8)
-    PQ8,
-    /// Product quantization with 4 bits per sub-space (PQ-4)
-    PQ4,
-    /// Binary quantization: 1 bit per dimension
-    Binary,
-}
-
-impl ZeroOverheadCollectionConfig {
-    /// Create config for FP32 collection
-    pub fn fp32(dimension: usize) -> Self {
-        Self {
-            dimension,
-            is_quantized: false,
-            quantization_method: None,
-        }
-    }
-
-    /// Create config for quantized collection
-    pub fn quantized(dimension: usize, method: QuantizationMethod) -> Self {
-        Self {
-            dimension,
-            is_quantized: true,
-            quantization_method: Some(method),
-        }
-    }
-
-    /// Calculate vector size in bytes based on configuration
-    pub fn vector_size_bytes(&self) -> usize {
-        if self.is_quantized {
-            match self.quantization_method {
-                Some(QuantizationMethod::INT8) => self.dimension,
-                Some(QuantizationMethod::PQ8) => self.dimension,
-                Some(QuantizationMethod::PQ4) => (self.dimension * 4).div_ceil(8),
-                Some(QuantizationMethod::Binary) => self.dimension.div_ceil(8),
-                None => self.dimension, // Default to 1 byte per dim
-            }
-        } else {
-            self.dimension * std::mem::size_of::<f32>()
-        }
-    }
-}
-
-/// Zero-overhead collection with shared configuration
-pub struct ZeroOverheadCollection {
-    /// The actual vectors - just raw data!
-    vectors: Vec<ZeroOverheadVector>,
-
-    /// ID to index mapping for O(1) lookup
-    id_index: DashMap<String, usize>,
-
-    /// Shared configuration (would come from collection cache in practice)
-    config: Arc<ZeroOverheadCollectionConfig>,
-}
-
-impl ZeroOverheadCollection {
-    /// Create new collection with configuration
-    pub fn new(config: ZeroOverheadCollectionConfig) -> Self {
-        Self {
-            vectors: Vec::new(),
-            id_index: DashMap::new(),
-            config: Arc::new(config),
-        }
-    }
-
-    /// Create with pre-allocated capacity
-    pub fn with_capacity(config: ZeroOverheadCollectionConfig, capacity: usize) -> Self {
-        Self {
-            vectors: Vec::with_capacity(capacity),
-            id_index: DashMap::with_capacity(capacity),
-            config: Arc::new(config),
-        }
-    }
-
-    /// Add FP32 vector
-    pub fn add_fp32(&mut self, id: String, vector: &[f32]) -> Result<()> {
-        if vector.len() != self.config.dimension {
-            return Err(anyhow!("Vector dimension mismatch"));
-        }
-
-        let zero_overhead = ZeroOverheadVector::from_fp32(&id, vector)?;
-
-        // Check if ID already exists and replace if so
-        if let Some(existing_index) = self.id_index.get(&id) {
-            self.vectors[*existing_index] = zero_overhead;
-        } else {
-            let index = self.vectors.len();
-            self.vectors.push(zero_overhead);
-            self.id_index.insert(id, index);
-        }
-        Ok(())
-    }
-
-    /// Add quantized vector
-    pub fn add_quantized(&mut self, id: String, quantized_vector: &[u8]) -> Result<()> {
-        let expected_size = self.config.vector_size_bytes();
-        if quantized_vector.len() != expected_size {
-            return Err(anyhow!("Quantized vector size mismatch"));
-        }
-
-        let zero_overhead = ZeroOverheadVector::from_quantized(&id, quantized_vector)?;
-
-        // Check if ID already exists and replace if so
-        if let Some(existing_index) = self.id_index.get(&id) {
-            self.vectors[*existing_index] = zero_overhead;
-        } else {
-            let index = self.vectors.len();
-            self.vectors.push(zero_overhead);
-            self.id_index.insert(id, index);
-        }
-        Ok(())
-    }
-
-    /// Get vector by ID with zero-copy access
-    #[inline]
-    pub fn get(&self, id: &str) -> Option<VectorView<'_>> {
-        self.id_index.get(id).map(|index| VectorView {
-            vector: &self.vectors[*index],
-            config: &self.config,
-        })
-    }
-
-    /// Get vector by index
-    #[inline]
-    pub fn by_index(&self, index: usize) -> Option<VectorView<'_>> {
-        self.vectors.get(index).map(|vector| VectorView {
-            vector,
-            config: &self.config,
-        })
-    }
-
-    /// Iterate over all vectors
-    pub fn iter(&self) -> impl Iterator<Item = VectorView<'_>> + '_ {
-        self.vectors.iter().map(|vector| VectorView {
-            vector,
-            config: &self.config,
-        })
-    }
-
-    /// Number of vectors
-    #[inline]
-    pub fn len(&self) -> usize {
-        self.vectors.len()
-    }
-
-    /// Check if empty
-    #[inline]
-    pub fn is_empty(&self) -> bool {
-        self.vectors.is_empty()
-    }
-
-    /// Total memory usage in bytes
-    pub fn memory_usage(&self) -> usize {
-        self.vectors.iter().map(|v| v.size_bytes()).sum()
-    }
-
-    /// Average bytes per vector
-    pub fn avg_bytes_per_vector(&self) -> f64 {
-        if self.is_empty() {
-            0.0
-        } else {
-            self.memory_usage() as f64 / self.len() as f64
-        }
-    }
-
-    /// Remove a vector by ID
-    pub fn remove(&mut self, id: &str) -> Option<ZeroOverheadVector> {
-        if let Some((_, index)) = self.id_index.remove(id) {
-            // Update indices for all vectors after the removed one
-            for mut entry in self.id_index.iter_mut() {
-                if *entry.value() > index {
-                    *entry.value_mut() -= 1;
-                }
-            }
-            // Remove and return the vector
-            Some(self.vectors.remove(index))
-        } else {
-            None
-        }
-    }
-
-    /// Get configuration
-    pub fn config(&self) -> &ZeroOverheadCollectionConfig {
-        &self.config
-    }
-}
-
-/// Zero-copy view of a vector with its configuration
-pub struct VectorView<'a> {
-    vector: &'a ZeroOverheadVector,
-    config: &'a ZeroOverheadCollectionConfig,
-}
-
-impl<'a> VectorView<'a> {
-    /// Get vector as FP32 slice (zero-copy)
-    #[inline]
-    pub fn as_f32(&self) -> Option<&[f32]> {
-        if !self.config.is_quantized {
-            Some(self.vector.as_f32(self.config.dimension))
-        } else {
-            None
-        }
-    }
-
-    /// Get vector as quantized bytes
-    #[inline]
-    pub fn as_quantized(&self) -> Option<&[u8]> {
-        if self.config.is_quantized {
-            Some(self.vector.as_quantized(self.config.vector_size_bytes()))
-        } else {
-            None
-        }
-    }
-
-    /// Get vector ID
-    #[inline]
-    pub fn id(&self) -> &str {
-        self.vector.id(self.config.vector_size_bytes())
-    }
-
-    /// Get raw vector for direct operations
-    #[inline]
-    pub fn raw(&self) -> &ZeroOverheadVector {
-        self.vector
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_zero_overhead_fp32() {
-        let config = ZeroOverheadCollectionConfig::fp32(3);
-        let mut collection = ZeroOverheadCollection::new(config);
-
-        collection
-            .add_fp32("vec1".to_string(), &[1.0, 2.0, 3.0])
-            .unwrap();
-        collection
-            .add_fp32("vec2".to_string(), &[4.0, 5.0, 6.0])
-            .unwrap();
-
-        assert_eq!(collection.len(), 2);
-
-        let view = collection.get("vec1").unwrap();
-        assert_eq!(view.id(), "vec1");
-        assert_eq!(view.as_f32().unwrap(), &[1.0, 2.0, 3.0]);
-
-        // Memory efficiency check
-        // 3 * 4 bytes (vector) + 4 bytes (id) = 16 bytes per vector
-        assert_eq!(collection.avg_bytes_per_vector(), 16.0);
-    }
-
-    #[test]
-    fn test_zero_overhead_quantized() {
-        let config = ZeroOverheadCollectionConfig::quantized(4, QuantizationMethod::INT8);
-        let mut collection = ZeroOverheadCollection::new(config);
-
-        let quantized = vec![128, 255, 0, 64];
-        collection
-            .add_quantized("q1".to_string(), &quantized)
-            .unwrap();
-
-        let view = collection.get("q1").unwrap();
-        assert_eq!(view.id(), "q1");
-        assert_eq!(view.as_quantized().unwrap(), &quantized[..]);
-
-        // 4 bytes (quantized) + 2 bytes (id) = 6 bytes
-        assert_eq!(collection.avg_bytes_per_vector(), 6.0);
-    }
-
-    #[test]
-    fn test_memory_savings() {
-        // Demonstrate memory savings for 384-dimensional vectors
-
-        // FP32: 384 * 4 = 1536 bytes for vector data
-        let fp32_config = ZeroOverheadCollectionConfig::fp32(384);
-        assert_eq!(fp32_config.vector_size_bytes(), 1536);
-
-        // INT8: 384 * 1 = 384 bytes (75% reduction)
-        let int8_config = ZeroOverheadCollectionConfig::quantized(384, QuantizationMethod::INT8);
-        assert_eq!(int8_config.vector_size_bytes(), 384);
-
-        // PQ4: 384 * 0.5 = 192 bytes (87.5% reduction)
-        let pq4_config = ZeroOverheadCollectionConfig::quantized(384, QuantizationMethod::PQ4);
-        assert_eq!(pq4_config.vector_size_bytes(), 192);
-
-        // Binary: 384 / 8 = 48 bytes (96.9% reduction)
-        let binary_config =
-            ZeroOverheadCollectionConfig::quantized(384, QuantizationMethod::Binary);
-        assert_eq!(binary_config.vector_size_bytes(), 48);
-    }
-}
+//! The zero-overhead vector storage types (`ZeroOverheadVector`,
+//! `ZeroOverheadCollection`, `CollectionConfig`, `QuantizationMethod`, …) now
+//! live in the `proximadb-index-storage` foundation crate (extracted from the
+//! root crate to shrink single-compile RSS — part of the root-crate
+//! decomposition track). This file preserves the
+//! `crate::index::axis::zero_overhead_vector::*` path for existing callers and
+//! is a pure forwarding shim. New code should depend on
+//! `proximadb_index_storage::zero_overhead_vector` directly.
+pub use proximadb_index_storage::zero_overhead_vector::*;
